order done!

+++ app/Http/Controllers/Admin/TetimonialController.php
@@ -4,20 +4,31 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Testimonial;
+use App\Traits\ImageTrait;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Storage;
 
 class TetimonialController extends Controller
 {
-       /**
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $testimonial = Testimonial::all();
+        $query = Testimonial::query();
+        $query->when('keyword', function ($sub) use ($request) {
+            $keyword = $request->keyword;
+            $sub->where(function ($q) use ($keyword) {
+                $q->where('name', 'LIKE', "%$keyword%");
+            });
+        });
+        $testimonial = $query->paginate(10);
+        if ($request->wantsJson()) {
+            return view('pages.dashboard.testimonial.pagination', compact('testimonial'))->render();
+        }
         return view('pages.dashboard.testimonial.index', compact('testimonial'));
     }
 
@@ -39,26 +50,42 @@ class TetimonialController extends Controller
      */
     public function store(Request $request)
     {
-        
-        
+
+
         $this->validate($request, [
             'foto' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
             'name' => 'required',
             'jabatan' => 'required',
             'deskripsi' => 'required'
         ]);
+        try {
+            $file_name = $request->foto->getClientOriginalName();
+            $foto = $request->foto->storeAs('public/testimonial', $file_name);
 
-        $file_name = $request->foto->getClientOriginalName();
-        $foto = $request->foto->storeAs('testimonial', $file_name);
+            $testimonial = Testimonial::create([
+                'foto' => $foto,
+                'name' => $request->name,
+                'jabatan' => $request->jabatan,
+                'deskripsi' => $request->deskripsi,
 
-        Testimonial::create([
-            'foto' => $foto,
-            'name' => $request->name,
-            'jabatan' => $request->jabatan,
-            'deskripsi' => $request->deskripsi,
-
-        ]);
-        return redirect()->back();
+            ]);
+            return response()->json([
+                'status' => true,
+                'message' => [
+                    'head' => 'Berhasil',
+                    'body' => "Testimonial $testimonial->name berhasil dibuat!"
+                ]
+            ], 201);
+        } catch (\Throwable $th) {
+            return response()->json([
+                'status' => false,
+                'message' => [
+                    'head' => 'Gagal',
+                    'body' =>
+                    json_encode($th)
+                ]
+            ], 500);
+        }
     }
 
     /**
@@ -80,7 +107,15 @@ class TetimonialController extends Controller
      */
     public function edit($id)
     {
-        //
+        $testimonial = Testimonial::find($id);
+        if (!$testimonial) return response()->json([
+            'status' => false,
+            'message' => [
+                'head' => 'Gagal',
+                'body' => "Testimonial dengan id $id tidak ditemukan."
+            ]
+        ], 404);
+        return response()->json(['data' => $testimonial], 200);
     }
 
     /**
@@ -98,11 +133,26 @@ class TetimonialController extends Controller
             'jabatan' => 'required',
             'deskripsi' => 'required'
         ]);
-
-        $file_name = $request->foto->getClientOriginalName();
-        $foto = $request->foto->storeAs('testimonial', $file_name);
+        $foto = "";
 
         $testimonial = Testimonial::find($id);
+
+        if (!$testimonial) return response()->json([
+            'status' => false,
+            'message' => [
+                'head' => 'Gagal',
+                'body' => "Testimonial dengan id $id tidak ditemukan."
+            ]
+        ], 404);
+
+        if ($request->foto) {
+            (new ImageTrait())->delete($testimonial->foto);
+            $file_name = $request->foto->getClientOriginalName();
+            $foto = $request->foto->storeAs('public/testimonial', $file_name);
+        } else {
+            $foto = $testimonial->foto;
+        }
+
         $testimonial->update([
             'foto' => $foto,
             'name' => $request->name,
@@ -110,7 +160,13 @@ class TetimonialController extends Controller
             'deskripsi' => $request->deskripsi,
 
         ]);
-        return redirect()->back();
+        return response()->json([
+            'status' => true,
+            'message' => [
+                'head' => 'Berhasil',
+                'body' => "Testimonial $testimonial->name berhasil diupdate!"
+            ]
+        ], 200);
     }
 
     /**
@@ -121,14 +177,33 @@ class TetimonialController extends Controller
      */
     public function destroy($id)
     {
-        $testimonial = Testimonial::find($id);
-        $this->removeImage($testimonial->foto);
+        try {
+            $testimonial = Testimonial::find($id);
+            if (!$testimonial) {
+                return $this->errorResponse("Not found", 404);
+            }
+            $this->removeImage($testimonial->foto);
 
-        if (Storage::delete($testimonial->foto)) {
-            $testimonial->delete();
+            if (Storage::delete($testimonial->foto)) {
+                $testimonial->delete();
+            }
+
+            return response()->json([
+                'status' => true,
+                'message' => [
+                    'head' => 'Berhasil',
+                    'body' => "Testimonial $testimonial->name berhasil dihapus!"
+                ]
+            ], 200);
+        } catch (\Throwable $th) {
+            return response()->json([
+                'status' => false,
+                'message' => [
+                    'head' => 'Gagal',
+                    'body' => $th->getMessage()
+                ]
+            ], 500);
         }
-
-        return redirect()->back();
     }
 
     public function removeImage($path)
